refactor(checker): split cross_file interface lowering under 2000-LOC limit
Adding the cross-file type-alias cycle guard pushed
`state/type_analysis/cross_file.rs` past the hard 2000-line file-size limit.
Move `lower_cross_file_interface_decl` and `merge_cross_file_heritage` (the
cross-file interface declaration lowering / heritage merge concern) into a new
`cross_file_interface_merge` module. Behavior-preserving: the methods keep the
same signatures and `CheckerState` impl; only `resolve_cross_file_heritage_type_arg`
is widened to `pub(crate)` so the moved heritage merge can still call it.
