+++ crates/tsz-checker/src/state/type_analysis/mod.rs
@@ -25,6 +25,7 @@ mod cross_file_direct_functions;
 mod cross_file_env_merge;
 mod cross_file_globals;
 mod cross_file_import_alias_pin;
+mod cross_file_interface_merge;
 pub(crate) mod cross_file_query_types;
 mod cross_file_residue;
 mod cross_file_shared_cache;
